Try reading .bib files from the filesystem when they aren’t in the files object
Previously, all bibliography files mentioned in the collections option were
assumed to have their content provided in metalsmith’s files object. Other
metalsmith plugins violate this assumption, for instance, to enable recompiling
only a subset of the source files.

This change implements a fallback: if a bibliography file is not present in the
files object, we then try to read the file directly from the filesystem. This
allows metalsmith-bibtex to coexist with plugins like metalsmith-watch.